# perf: Return `Cow<Signature>` where possible