--- a/hugr-core/src/ops/constant.rs
+++ b/hugr-core/src/ops/constant.rs
@@ -2,6 +2,7 @@
 
 mod custom;
 
+use std::borrow::Cow;
 use std::collections::hash_map::DefaultHasher; // Moves into std::hash in Rust 1.76.
 use std::hash::{Hash, Hasher};
 
@@ -348,12 +349,15 @@ pub enum ConstTypeError {
 }
 
 /// Hugrs (even functions) inside Consts must be monomorphic
-fn mono_fn_type(h: &Hugr) -> Result<Signature, ConstTypeError> {
+fn mono_fn_type(h: &Hugr) -> Result<Cow<'_, Signature>, ConstTypeError> {
     let err = || ConstTypeError::NotMonomorphicFunction {
         hugr_root_type: h.root_type().clone(),
     };
     if let Some(pf) = h.poly_func_type() {
-        return pf.try_into().map_err(|_| err());
+        match pf.try_into() {
+            Ok(sig) => return Ok(Cow::Owned(sig)),
+            Err(_) => return Err(err()),
+        };
     }
 
     h.inner_function_type().ok_or_else(err)
@@ -367,7 +371,7 @@ impl Value {
             Self::Sum(Sum { sum_type, .. }) => sum_type.clone().into(),
             Self::Function { hugr } => {
                 let func_type = mono_fn_type(hugr).unwrap_or_else(|e| panic!("{}", e));
-                Type::new_function(func_type)
+                Type::new_function(func_type.into_owned())
             }
         }
     }