--- a/hugr-core/src/std_extensions/arithmetic/int_ops.rs
+++ b/hugr-core/src/std_extensions/arithmetic/int_ops.rs
@@ -384,24 +384,27 @@ mod test {
                 .with_two_log_widths(3, 4)
                 .to_extension_op()
                 .unwrap()
-                .signature(),
-            Signature::new(int_type(3), int_type(4)).with_extension_delta(EXTENSION_ID)
+                .signature()
+                .as_ref(),
+            &Signature::new(int_type(3), int_type(4)).with_extension_delta(EXTENSION_ID)
         );
         assert_eq!(
             IntOpDef::iwiden_s
                 .with_two_log_widths(3, 3)
                 .to_extension_op()
                 .unwrap()
-                .signature(),
-            Signature::new_endo(int_type(3)).with_extension_delta(EXTENSION_ID)
+                .signature()
+                .as_ref(),
+            &Signature::new_endo(int_type(3)).with_extension_delta(EXTENSION_ID)
         );
         assert_eq!(
             IntOpDef::inarrow_s
                 .with_two_log_widths(3, 3)
                 .to_extension_op()
                 .unwrap()
-                .signature(),
-            Signature::new(int_type(3), sum_ty_with_err(int_type(3)))
+                .signature()
+                .as_ref(),
+            &Signature::new(int_type(3), sum_ty_with_err(int_type(3)))
                 .with_extension_delta(EXTENSION_ID)
         );
         assert!(
@@ -417,8 +420,9 @@ mod test {
                 .with_two_log_widths(2, 1)
                 .to_extension_op()
                 .unwrap()
-                .signature(),
-            Signature::new(int_type(2), sum_ty_with_err(int_type(1)))
+                .signature()
+                .as_ref(),
+            &Signature::new(int_type(2), sum_ty_with_err(int_type(1)))
                 .with_extension_delta(EXTENSION_ID)
         );