--- a/hugr-llvm/src/utils/inline_constant_functions.rs
+++ b/hugr-llvm/src/utils/inline_constant_functions.rs
@@ -66,6 +66,7 @@ fn inline_constant_functions_impl(
                     "Constant function hugr has no inner_func_type: {}",
                     konst_n.index()
                 ))?
+                .into_owned()
                 .into();
             let func_defn = FuncDefn {
                 name: const_fn_name(konst_n),