```
group                                    cows                                   serialization
-----                                    ----                                   -------------
circuit_roundtrip/capnp/0                1.02      5.3±0.52µs        ? ?/sec    1.00      5.2±0.09µs        ? ?/sec
circuit_roundtrip/capnp/1                1.00     12.5±0.74µs        ? ?/sec    1.14     14.2±0.16µs        ? ?/sec
circuit_roundtrip/capnp/10               1.00     25.9±0.39µs        ? ?/sec    1.31     34.0±0.30µs        ? ?/sec
circuit_roundtrip/capnp/100              1.00    155.8±2.01µs        ? ?/sec    1.47    229.6±2.46µs        ? ?/sec
circuit_roundtrip/capnp/1000             1.00  1483.7±64.06µs        ? ?/sec    1.43      2.1±0.04ms        ? ?/sec
circuit_roundtrip/capnp/10000            1.00     15.9±0.26ms        ? ?/sec    1.49     23.7±1.67ms        ? ?/sec
circuit_roundtrip/capnp/100000           1.00    192.8±2.26ms        ? ?/sec    1.37    264.3±5.44ms        ? ?/sec
circuit_roundtrip/capnp/200000           1.00    411.5±4.77ms        ? ?/sec    1.36   560.4±15.52ms        ? ?/sec
circuit_roundtrip/capnp/300000           1.00    648.9±5.40ms        ? ?/sec    1.34   871.5±10.63ms        ? ?/sec
circuit_roundtrip/json/0                 1.00      8.0±0.18µs        ? ?/sec    1.00      8.0±0.14µs        ? ?/sec
circuit_roundtrip/json/1                 1.00     19.0±0.27µs        ? ?/sec    1.16     22.1±0.33µs        ? ?/sec
circuit_roundtrip/json/10                1.00     50.0±0.68µs        ? ?/sec    1.26     63.0±0.94µs        ? ?/sec
circuit_roundtrip/json/100               1.00    342.5±4.81µs        ? ?/sec    1.30    446.4±4.61µs        ? ?/sec
circuit_roundtrip/json/1000              1.00      3.4±0.03ms        ? ?/sec    1.31      4.5±0.35ms        ? ?/sec
circuit_roundtrip/json/10000             1.00     37.2±0.73ms        ? ?/sec    1.28     47.5±0.45ms        ? ?/sec
circuit_roundtrip/json/100000            1.00    399.3±8.46ms        ? ?/sec    1.27   507.7±16.13ms        ? ?/sec
circuit_roundtrip/json/1000000           1.00       4.5±0.26s        ? ?/sec    1.20       5.4±0.06s        ? ?/sec
``` 

Closes #1741