@aborgna-q 2024-12-10
I had to add a small workaround for https://github.com/CQCL/hugr/issues/1758, since now we actually check the extension_reqs that come from the serialized hugr.

I'll fix it in another PR